+++ CHANGELOG.md
@@ -5,6 +5,14 @@ All notable changes to SCOPE-XR will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.3.4] - 2026-07-31
+
+### Changed
+- FIX: Fixed various linting and typing errors
+- FIX: Hough detection now gracefully fails
+- FIX: psf_runner now correctly applies the image cropping around the circle
+
+
 ## [1.3.3] - 2026-07-27
 
 ### Changed